refactor: organize api utils
