+++ frontend/src/api/users.ts
@@ -1,4 +1,4 @@
-import { fetchApi, routeBuilder } from "./util";
+import { fetchApi, routeBuilder } from "./utils";
 
 const usersRoute = routeBuilder("/api/users");
